pyproject.toml
[project]
name = "vvo-display"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
